refactor: reduce build size and allow local dev
