+++ package.json
@@ -61,6 +61,7 @@
     "prettier": "^2.8.8",
     "rimraf": "^5.0.1",
     "supertest": "^6.3.3",
+    "terser-webpack-plugin": "^5.3.9",
     "ts-jest": "29.1.0",
     "ts-loader": "^9.4.3",
     "ts-node": "10.9.1",
